- user is logged in redirect to browse and vice versa - bugfix
