+++ src/components/Body.js
@@ -1,14 +1,14 @@
-import React, { useEffect } from "react";
+import React from "react";
 import Login from "./Login";
 import Browse from "./Browse";
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
-import { auth } from "../utils/firebase";
-import { onAuthStateChanged } from "firebase/auth";
-import { useDispatch } from "react-redux";
-import { addUser, removeUser } from "../utils/userSlice";
+// import { auth } from "../utils/firebase";
+// import { onAuthStateChanged } from "firebase/auth";
+// import { useDispatch } from "react-redux";
+// import { addUser, removeUser } from "../utils/userSlice";
 
 const Body = () => {
-  const dispatch = useDispatch();
+  // const dispatch = useDispatch();
 
   const appRouter = createBrowserRouter([
     {
@@ -21,26 +21,27 @@ const Body = () => {
     },
   ]);
 
-  useEffect(() => {
-    onAuthStateChanged(auth, (user) => {
-      if (user) {
-        // User is signed in, see docs for a list of available properties
-        // https://firebase.google.com/docs/reference/js/auth.user
-        const { uid, email, displayName, photoURL } = user;
-        dispatch(
-          addUser({
-            uid: uid,
-            email: email,
-            displayName: displayName,
-            photoURL: photoURL,
-          })
-        );
-      } else {
-        // User is signed out
-        dispatch(removeUser());
-      }
-    });
-  }, []);
+  // useEffect(() => {
+  //   onAuthStateChanged(auth, (user) => {
+  //     if (user) {
+  //       // User is signed in, see docs for a list of available properties
+  //       // https://firebase.google.com/docs/reference/js/auth.user
+  //       const { uid, email, displayName, photoURL } = user;
+  //       dispatch(
+  //         addUser({
+  //           uid: uid,
+  //           email: email,
+  //           displayName: displayName,
+  //           photoURL: photoURL,
+  //         })
+  //       );
+
+  //     } else {
+  //       // User is signed out
+  //       dispatch(removeUser());
+  //     }
+  //   });
+  // }, []);
 
   return (
     <div>
